fix taking the absolute value of unsigned type

+++ src/aptex-defs.h
@@ -46,7 +46,7 @@
 
 // predefined macro
 // #define abs(x)      ((integer)(x) >= 0 ? (integer)(x) : (integer)-(x))
-#define abs(x)      labs(x)
+#define abs(x)      labs((integer)x)
 #define chr(x)      (x)
 #define odd(x)      ((x) % 2)
 #define round(x)    aptex_utils_round((real) (x))
